Revamp Syncing strategy.
The previous syncing strategy was a bit intrasparent and confusing. This
is an attempt to clean it up. Changes 1-3 are concerned
with changes in accounts.py. The rest is implemented in folder/Base.py.

1) Do away with the previous different code paths depending on
whether there is a LocalStatus file or not (the isnewfolder() test). We
always use the same strategy now.

This strategy is simply:
a) Sync remote to local folder first
b) Sync local to remote

Where each sync implies a 4 pass strategy as before (explained below).

2) Don't delete messages on LOCAL which don't exist on REMOTE right at
the beginning. This prevented us from keeping local messages rather than
redownloading everything once LocalStatus got corrupted or deleted.

3) No need to separately sync the statusfolder anymore, we update that
one simultanously with the destfolders...

3) Simplified the sync functions by only taking one destdir, we
never used more anyway. This makes the code better to understand.

4) Added plenty of code comments while I was going through to make sure it
is easy to understand.

A folder sync strategy is mostly like it was before:
        Pass1: Transfer new local messages
               Upload msg with negative/no UIDs to dstfolder. dstfolder
               should assign that message a new UID. Update statusfolder.
        Pass2: Copy existing messages
               Copy messages in self, but not statusfolder to dstfolder
               if not already in dstfolder. Update statusfolder.
        Pass3: Remove deleted messages
               Get all UIDS in statusfolder but not self. These are
               messages that we have locally deleted. Delete those from
               dstfolder and statusfolder.
        Pass4: Synchronize flag changes
               Compare flags in self with those in statusfolder. If msg
               has a valid UID and exists on dstfolder (has not e.g.
               been deleted there), sync the flag change to dstfolder
               and statusfolder.

The user visible implications of this change should be limited:
 - Blowing away LocalStatus will not require you to redownload ALL of
 your mails if you still have the local Maildir. It will simply recreate
 LocalStatus.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>
Signed-off-by: Nicolas Sebrecht <nicolas.s-dev@laposte.net>